+++ platform/mellanox/mlnx-platform-api/sonic_platform/thermal_updater.py
@@ -31,67 +31,11 @@ TC_CONFIG_FILE = '/run/hw-management/config/tc_config.json'
 logger = logger.Logger('thermal-updater')
 
 
-class TimerEvent:
-    def __init__(self, interval, cb, repeat):
-        self.interval = interval
-        self._cb = cb
-        self.repeat = repeat
-
-    def execute(self):
-        self._cb()
-
-
-class Timer(threading.Thread):
-    def __init__(self):
-        super(Timer, self).__init__()
-        self._timestamp_queue = queue.PriorityQueue()
-        self._wait_event = threading.Event()
-        self._stop_event = threading.Event()
-        self._min_timestamp = None
-        self._status = {}
-
-    def schedule(self, interval, cb, repeat=True, run_now=True):
-        timer_event = TimerEvent(interval, cb, repeat)
-        self.add_timer_event(timer_event, run_now)
-
-    def add_timer_event(self, timer_event, run_now=True):
-        timestamp = time.time()
-        if not run_now:
-            timestamp += timer_event.interval
-
-        self._timestamp_queue.put_nowait((timestamp, timer_event))
-        if self._min_timestamp is not None and timestamp < self._min_timestamp:
-            self._wait_event.set()
-
-    def stop(self):
-        was_alive = self.is_alive()
-        self._wait_event.set()
-        self._stop_event.set()
-        if was_alive:
-            self.join()
-
-    def run(self):
-        while not self._stop_event.is_set():
-            now = time.time()
-            item = self._timestamp_queue.get()
-            self._min_timestamp = item[0]
-            if self._min_timestamp > now:
-                self._wait_event.wait(self._min_timestamp - now)
-                self._wait_event.clear()
-                self._timestamp_queue.put(item)
-                continue
-
-            timer_event = item[1]
-            timer_event.execute()
-            if timer_event.repeat:
-                self.add_timer_event(timer_event, False)
-
-
 class ThermalUpdater:
     def __init__(self, sfp_list):
         self._sfp_list = sfp_list
         self._sfp_status = {}
-        self._timer = Timer()
+        self._timer = utils.Timer()
 
     def load_tc_config(self):
         asic_poll_interval = 1
@@ -159,7 +103,7 @@ class ThermalUpdater:
         temperature = utils.read_int_from_file('/sys/module/sx_core/asic0/temperature/input', default=None)
         return temperature * ASIC_TEMPERATURE_SCALE if temperature is not None else ASIC_DEFAULT_TEMP_WARNNING_THRESHOLD
 
-    def get_asic_temp_waning_threashold(self):
+    def get_asic_temp_warning_threashold(self):
         emergency = utils.read_int_from_file('/sys/module/sx_core/asic0/temperature/emergency', default=None, log_func=None)
         return emergency * ASIC_TEMPERATURE_SCALE if emergency is not None else ASIC_DEFAULT_TEMP_WARNNING_THRESHOLD
 
@@ -214,7 +158,7 @@ class ThermalUpdater:
     def update_asic(self):
         try:
             self.set_thermal_data(0, 'input', self.get_asic_temp())
-            self.set_thermal_data(0, 'emergency', self.get_asic_temp_waning_threashold())
+            self.set_thermal_data(0, 'emergency', self.get_asic_temp_warning_threashold())
             self.set_thermal_data(0, 'crit', self.get_asic_temp_critical_threashold())
         except Exception as e:
             logger.log_error('Failed to update ASIC thermal data - {e}')
